requirements.txt
streamlit
torch
torchaudio
openai-whisper
transformers
moviepy
imageio
imageio-ffmpeg
numpy